fix: add list-reorgs command
