+++ applications/tari_base_node/src/commands/command/list_connections.rs
@@ -14,7 +14,7 @@ pub struct Args {}
 
 #[async_trait]
 impl HandleCommand<Args> for CommandContext {
-    async fn handle_command(&mut self, args: Args) -> Result<(), Error> {
+    async fn handle_command(&mut self, _: Args) -> Result<(), Error> {
         self.list_connections().await
     }
 }
