+++ applications/tari_base_node/src/commands/command/unban_all_peers.rs
@@ -3,10 +3,8 @@ use async_trait::async_trait;
 use chrono::Utc;
 use clap::Parser;
 use tari_comms::peer_manager::{PeerFeatures, PeerQuery};
-use tari_core::base_node::state_machine_service::states::PeerMetadata;
 
 use super::{CommandContext, HandleCommand};
-use crate::{table::Table, utils::format_duration_basic};
 
 /// Unbans all peers
 #[derive(Debug, Parser)]
@@ -14,7 +12,7 @@ pub struct Args {}
 
 #[async_trait]
 impl HandleCommand<Args> for CommandContext {
-    async fn handle_command(&mut self, args: Args) -> Result<(), Error> {
+    async fn handle_command(&mut self, _: Args) -> Result<(), Error> {
         self.unban_all_peers().await
     }
 }
