+++ applications/tari_base_node/src/commands/command_handler.rs
@@ -48,12 +48,7 @@ use tari_comms::{
 };
 use tari_comms_dht::{envelope::NodeDestination, DhtDiscoveryRequester, MetricsCollectorHandle};
 use tari_core::{
-    base_node::{
-        comms_interface::BlockEvent,
-        state_machine_service::states::{PeerMetadata, StatusInfo},
-        LocalNodeCommsInterface,
-    },
-    blocks::{BlockHeader, ChainHeader},
+    base_node::{comms_interface::BlockEvent, state_machine_service::states::StatusInfo, LocalNodeCommsInterface},
     chain_storage::{async_db::AsyncBlockchainDb, LMDBDatabase},
     consensus::ConsensusManager,
     mempool::service::LocalMempoolService,
@@ -66,14 +61,9 @@ use tari_p2p::{
 };
 use tari_utilities::{hex::Hex, message_format::MessageFormat, Hashable};
 use thiserror::Error;
-use tokio::{
-    fs::File,
-    io::AsyncWriteExt,
-    sync::{broadcast, watch},
-};
+use tokio::{fs::File, io::AsyncWriteExt, sync::watch};
 
-use super::status_line::StatusLine;
-use crate::{builder::BaseNodeContext, table::Table, utils::format_duration_basic, LOG_TARGET};
+use crate::{builder::BaseNodeContext, table::Table};
 
 #[derive(Debug, Display, EnumString)]
 #[strum(serialize_all = "lowercase")]
@@ -494,32 +484,6 @@ impl CommandHandler {
             }
         }
         table.print_stdout();
-        Ok(())
-    }
-
-    pub fn list_reorgs(&self) -> Result<(), Error> {
-        if !self.config.blockchain_track_reorgs {
-            // TODO: Return error/report
-            println!(
-                "Reorg tracking is turned off. Add `track_reorgs = true` to the [base_node] section of your config to \
-                 turn it on."
-            );
-        } else {
-            let reorgs = self.blockchain_db.inner().fetch_all_reorgs()?;
-            let mut table = Table::new();
-            table.set_titles(vec!["#", "New Tip", "Prev Tip", "Depth", "Timestamp"]);
-
-            for (i, reorg) in reorgs.iter().enumerate() {
-                table.add_row(row![
-                    i + 1,
-                    format!("#{} ({})", reorg.new_height, reorg.new_hash.to_hex()),
-                    format!("#{} ({})", reorg.prev_height, reorg.prev_hash.to_hex()),
-                    format!("{} added, {} removed", reorg.num_blocks_added, reorg.num_blocks_removed),
-                    reorg.local_time
-                ]);
-            }
-            table.enable_row_count().print_stdout();
-        }
         Ok(())
     }
 }
