+++ applications/tari_base_node/src/commands/parser.rs
@@ -42,7 +42,6 @@ pub enum BaseNodeCommand {
     RewindBlockchain,
     PeriodStats,
     HeaderStats,
-    ListReorgs,
     DiscoverPeer,
     GetBlock,
     SearchUtxo,
